% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenoisy_asf.R
\name{flipout}
\alias{flipout}
\title{Change outcome column values.}
\usage{
flipout(data, outcome, proportion)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{outcome}{Character; outcome column (factor) name.}

\item{proportion}{A numeric that determines a proportion of rows in \code{data}
for which outcome is to be changed.}
}
\value{
A \code{data.frame}
}
\description{
Create a noisy data set by changing outcome column (factor) values for a
proportion of rows of a data set.
}
\details{
Given a data frame \code{data} and the name of an outcome variable/factor
\code{outcome}, \code{flipout()} changes the outcome value for a proportion of rows
in \code{data}, the number of which is geiven by \code{proportion}. If \code{proportion}
is not a (vulgar) fraction of \code{nrow(data)}, number of rows to be
manipulated is determined as \code{round(proportion * nrow(data))}.
}
