% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscorrect.R
\name{is_correct}
\alias{is_correct}
\title{is_correct}
\usage{
is_correct(a, b)
}
\arguments{
\item{a}{a CNA solution}

\item{b}{a CNA solution}
}
\value{
Logical
}
\description{
Wrapper for \code{is.submodel} that changes the handling of \code{NA}s and \code{NULL}s
}
\details{
A wrapper of \code{is.submodel} for use in checking the correctness
of solutions created and stored programmatically (such that empty solutions
may get converted to \code{NA}s or \code{NULL}s). Returns \code{TRUE} if first
argument is \code{NA} or \code{NULL}.
}
