% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenoisy_asf.R
\name{makenoisy_asf}
\alias{makenoisy_asf}
\title{Make a noisy data set from an \emph{asf} while controlling outcome prevalence.}
\usage{
makenoisy_asf(
  model = randomAsf(6),
  data = ct2df(selectCases(model)),
  outcome = setNames(list(1), rhs(model)),
  prevalence = 0.5,
  noiselevel = 2/N,
  N = nrow(data)
)
}
\arguments{
\item{model}{A character string that expresses a \code{cna} model.}

\item{data}{A data.frame; the data set to be manipulated.}

\item{outcome}{Named, unit length list; the outcome variable/factor name and value.}

\item{prevalence}{Numeric between zero and one.}

\item{noiselevel}{Numeric between zero and one.}

\item{N}{Integer.}
}
\value{
\code{data.frame}
}
\description{
Make a noisy data set from an \emph{asf} while controlling outcome prevalence.
}
\details{
\code{makenoisy_asf()} generates noisy data sets while allowing the user
to control outcome prevalence. Given an \emph{asf}, a data set will be created
where the outcome prevalence equals \code{prevalence}, number of rows equals
\code{N}, and proportion of rows that do not conform to the input \emph{asf} (i.e. the
proportion of noise to ideal data) equals \code{noiselevel}.
}
\examples{
makenoisy_asf("A+B<->C", prevalence = 0.25, noiselevel = 0.2, N = 10)
}
