% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_compliant_noisify.R
\name{prevalence_compliant_noisify}
\alias{prevalence_compliant_noisify}
\title{Introduce noise to data set while keeping outcome prevalence fixed.}
\usage{
prevalence_compliant_noisify(model, data, outcome, noiselevel)
}
\arguments{
\item{model}{}

\item{data}{}

\item{outcome}{}

\item{noiselevel}{}
}
\description{
Introduce noise to data set while keeping outcome prevalence fixed.
}
