% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDat.R
\name{randomDat}
\alias{randomDat}
\title{randomDat}
\usage{
randomDat(
  x = 5,
  condtype = c("csf", "asf"),
  type = c("cs", "mv", "fs"),
  mvlevels = if (type == "mv") c(0L:3L) else NULL,
  ...
)
}
\arguments{
\item{x}{A \code{configTable}, a data frame, a matrix, an integer, a list
specifying the factors' value ranges, or a character vector featuring all
admissible factor values}

\item{condtype}{Character vector determining if the data-generating structure
(DGS) can be a csf or not, defaults to csf.}

\item{type}{Character vector; determines the type of data generated}

\item{mvlevels}{Integer vector of length two. Determines the range of
permissible factor values when creating multi-value data. Only used if
\code{type = "mv"} and admissible values are not implicitly determined by
\code{x}}

\item{...}{Additional arguments passed to \code{randomConds},
\code{selectCases}, and \code{makefuzzy}}
}
\value{
A data frame
}
\description{
Simulate a data set from a randomly generated CNA solution in
one go
}
\details{
Primarily a more flexible replacement for a snippet
\code{ct2df(selectCases(randomConds(x)))}. By default produces a data set
over five crisp-set factors from a randomly generated csf. Other types of
data can be produced by changing the value of \code{type} to "mv" or "fs".
Abuses \code{...} to pass additional arguments to randomConds, selectCases,
and makeFuzzy (if \code{type = "fs"}). Only valid arguments are passed,
will not break if given unused arguments in \code{...}. The user may
control the creation of fuzzy set data by passing an explicit
\code{fuzzvalues} argument in \code{...}, otherwise makeFuzzy default
values are used. Note that argument \code{x} accepts any type of value that
can be passed to \code{full.ct}, but the underlying procedure is the same
regardless. The only purpose of assigning other than integer value to
\code{x} is to control the column names in the created data and admissible
factor values in mv data (otherwise controlled by the argument
\code{mvlevels}), if \code{x} is given a multi-value data set.
}
\examples{
randomDat(5, type = "mv")
}
