% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solgen.R
\name{solgen}
\alias{solgen}
\title{solgen}
\usage{
solgen(
  x = 5,
  range = c(0.8, 0.6),
  gran = 0.1,
  forceoutcome = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Either an integer greater or equal to four, data frame of binary
crisp set data, or a configTable object, to be passed to \code{randomDat}.
Determines the number, names, and type (crisp-set, multi-value or fuzzy) of
the factors in the data set used for the analyses, and consequently the
type of the solutions that are generated}

\item{range}{Numeric vector of length two, determining the upper and lower
bound of both consistency and coverage thresholds used in the reanalyses}

\item{gran}{A numeric greater than zero and less than one, determining
the granularity of variation in consistency and coverage thresholds between
the reanalyses.}

\item{forceoutcome}{Logical; if \code{TRUE}, one of the outcomes in the DGS
that returns the analyzed data is passed to CNA as the argument
\code{ordering}. Might increase the probability that the returned models
exhibit submodel-relations. Overriden by an explicit ordering argument
given in \code{...}}

\item{verbose}{Logical; if \code{TRUE}, returns a data frame including all
details returned by CNA}

\item{...}{Additional arguments passed to \code{randomDat} and \code{cna}}
}
\value{
A vector or data.frame of CNA solutions with extra attribute
"target" containing the DGS from which the data were simulated.
}
\description{
Generate CNA solutions likely to exhibit submodel-relations by repeatedly
analysing a data set returned by a randomly generated model while using
different consistency and coverage thresholds in each reanalysis of the data.
}
\details{
Primary use-case is to quickly generate a throwaway set of solutions
that can be used to test other functions that exploit the presence of
submodel-relations in a set of solutions to asses correctness, robustness,
completeness etc. To that end, this function provides a shortcut for
producing a number of solutions fitted to the same underlying data using
different consistency and coverage thresholds. First a
data set is generated by \code{randomDat(x, ...)}. This data is then
repeatedly analyzed with \code{cna}, using every logically possible
combination of consistency and coverage thresholds determined by arguments
\code{range} and \code{gran}.  Checks the arguments given in \code{...},
only valid arguments are passed further, i.e. will not break if given
unused arguments. Note that if the value of \code{x} is a fuzzy set
configTable, the procedure used is identical to crisp set data. The only
practical purpose of passing fs data is to ensure that the created
solutions use the same factor names. If this is not specifically desired,
it is probably quicker to assign \code{x} an integer value equal to the
desired maximum number of factors in the solutions.
}
