% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDat.R
\name{randomDat}
\alias{randomDat}
\title{randomDat}
\usage{
randomDat(x = 5, condtype = c("csf", "asf"), fuzzy = FALSE, ...)
}
\arguments{
\item{x}{Positive scalar integer, data.frame of binary crisp set data, or a
configTable object, determining the number of factors in the created data
as well as their type}

\item{condtype}{Character vector determining if the data-generating structure
(DGS) can be a csf or not, defaults to csf. If "csf", \code{randomCsf} is
used to generate the DGS, otherwise \code{randomAsf}}

\item{fuzzy}{Logical; if \code{TRUE}, fuzzy-set data will be generated}

\item{...}{Additional arguments passed to \code{randomConds},
\code{selectCases}, and \code{makefuzzy}}
}
\description{
Simulate a data set from a randomly generated CNA solution in
one go
}
\details{
Primarily a replacement for a snippet
\code{ct2df(selectCases(randomConds(x)))}. By default produces a data set
over five crisp-set factors from a randomly generated csf. Abuses
\code{...} to pass additional arguments to randomConds, selectCases, and
makeFuzzy (if \code{fuzzy = TRUE}). Only valid arguments are passed, will
not brake if given unused arguments. The user may control the creation of
fuzzy set data by passing an explicit \code{fuzzvalues} argument in
\code{...}, otherwise makeFuzzy default values are used.
}
